Rework error handling and protocol update
This pull request deals with two issues that are dependent on each
other.

Error handling is reworked so it no longer uses gRPC metadata but
in-built gRPC error handling support via exceptions. gRPC submodule was
updated to reflect this change.

Proto protocol is updated. Appliance and Image sctructures was changed
and changes are described in
https://github.com/the-cloudkeeper-project/cloudkeeper-proto/pull/3.
Also method UpdateAppliance was added and this change is described in
https://github.com/the-cloudkeeper-project/cloudkeeper-proto/pull/4

Close #37